Update some apis
