--- frontend/src/util/userContext.tsx
@@ -1,7 +1,0 @@
-import { createContext } from "react";
-
-const UserContext = createContext({
-	uid: "",
-})
-
-export default UserContext;
